pyproject.toml
[project]
name = "videoprocessingpipeline"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "clickhouse-connect>=0.10.0",
    "lancedb>=0.26.1",
    "mypy>=1.19.1",
    "numpy>=2.4.0",
    "opencv-python>=4.11.0.86",
    "python-dotenv>=1.2.1",
    "ruff>=0.14.10",
    "temporalio>=1.21.1",
]

[dependency-groups]
dev = [
    "mypy>=1.19.1",
    "ruff>=0.14.10",
]
